requirements.txt
ultralytics 
opencv-python
langchain
python-dotenv
whisper
PyAudio
wave
SpeechRecognition
gradio
huggingface_hub
transformers
langchain-groq
openai-whisper
pyttsx3
streamlit